phase-IV/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
pywebpush>=1.14.0
httpx>=0.26.0
itsdangerous>=2.0.0
openai>=1.0.0
cohere>=5.0.0
authlib>=1.3.0